Computer technology
COMPUTER TECHNOLOGY
The Computer is one of the first technologies that made it easier for humans to record and reference their previous actions and thoughts. That technology were later improved upon with the development of the printing press. The computer is a machine  instructed to carry out and operations automatically. Modern computers have the ability and follow generalized sets of operations. There program enable to computers perform and extremely wide range of tasks. A "complete" computer including the hardware, the operating system, required and used for "full" operation can be referred to as a computer system. This is term may as well be used for a group of computers that are connected and work together, in particular, a computer network.
Technology Write for Us has also brought about efficiency and quality and the manufacturing sector! Technological advancement has reduced risk. Development is closely related to technology and stage of development the human and machine being has arrived could has been possible without the advancement in technology.
We can modernize our industry so life becomes easier for us generations through the developed technologies.
After successfully demonstrating a proof-of-concept prototype in 1939, Professor John Vincent Atanasoff receives funds to build a full-scale machine at Iowa State College (now University). The machine was designed and built by Atanasoff and graduate student Clifford Berry between 1939 and 1942. The ABC was at the center of a patent dispute related to the invention of the computer, which was resolved in 1973 when it was shown that ENIAC co-designer John Mauchly had seen the ABC shortly after it became functional.
The legal result was a landmark: Atanasoff was declared the originator of several basic computer ideas, but the computer as a concept was declared un-patentable and thus freely open to all. A full-scale working replica of the ABC was completed in 1997, proving that the ABC machine functioned as Atanasoff had claimed. The replica is currently on display at the Computer History Museum.
The US Army asked Bell Laboratories to design a machine to assist in testing its M-9 gun director, a type of analog computer that aims large guns to their targets. Mathematician George Stibitz recommends using a relay-based calculator for the project. The result was the Relay Interpolator, later called the Bell Labs Model II. The Relay Interpolator used 440 relays, and since it was programmable by paper tape, was used for other applications following the war.